// web/src/api/models/personResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface PersonResponse {
    id?: string;
    displayName?: string;
    firstName?: string;
    lastName?: string;
    birthDate?: string;
    gender?: string;
    createdAt?: string;
    activeAccountLinked?: boolean;
}